cleanup(storage): split object stream files
It was getting hard to find things in a large file containing both
`ObjectReadStreambuf` and `ObjectWriteStreambuf`.
