+++ google/cloud/storage/google_cloud_cpp_storage.bzl
@@ -23,6 +23,7 @@ google_cloud_cpp_storage_hdrs = [
     "client.h",
     "client_options.h",
     "download_options.h",
+    "hash_mismatch_error.h",
     "hashing_options.h",
     "hmac_key_metadata.h",
     "iam_policy.h",
@@ -72,8 +73,9 @@ google_cloud_cpp_storage_hdrs = [
     "internal/object_acl_requests.h",
     "internal/object_metadata_parser.h",
     "internal/object_read_source.h",
+    "internal/object_read_streambuf.h",
     "internal/object_requests.h",
-    "internal/object_streambuf.h",
+    "internal/object_write_streambuf.h",
     "internal/openssl_util.h",
     "internal/parameter_pack_validation.h",
     "internal/patch_builder.h",
@@ -111,8 +113,10 @@ google_cloud_cpp_storage_hdrs = [
     "oauth2/service_account_credentials.h",
     "object_access_control.h",
     "object_metadata.h",
+    "object_read_stream.h",
     "object_rewriter.h",
     "object_stream.h",
+    "object_write_stream.h",
     "options.h",
     "override_default_project.h",
     "parallel_upload.h",
@@ -175,8 +179,9 @@ google_cloud_cpp_storage_srcs = [
     "internal/object_access_control_parser.cc",
     "internal/object_acl_requests.cc",
     "internal/object_metadata_parser.cc",
+    "internal/object_read_streambuf.cc",
     "internal/object_requests.cc",
-    "internal/object_streambuf.cc",
+    "internal/object_write_streambuf.cc",
     "internal/openssl_util.cc",
     "internal/patch_builder.cc",
     "internal/policy_document_request.cc",
@@ -206,8 +211,9 @@ google_cloud_cpp_storage_srcs = [
     "oauth2/service_account_credentials.cc",
     "object_access_control.cc",
     "object_metadata.cc",
+    "object_read_stream.cc",
     "object_rewriter.cc",
-    "object_stream.cc",
+    "object_write_stream.cc",
     "parallel_upload.cc",
     "policy_document.cc",
     "service_account.cc",
