+++ google/cloud/storage/storage_client_unit_tests.bzl
@@ -66,8 +66,9 @@ storage_client_unit_tests = [
     "internal/metadata_parser_test.cc",
     "internal/notification_requests_test.cc",
     "internal/object_acl_requests_test.cc",
+    "internal/object_read_streambuf_test.cc",
     "internal/object_requests_test.cc",
-    "internal/object_streambuf_test.cc",
+    "internal/object_write_streambuf_test.cc",
     "internal/openssl_util_test.cc",
     "internal/parameter_pack_validation_test.cc",
     "internal/patch_builder_test.cc",
